Move SelectedAppTopBar to ui
