+++ core/ui/src/main/java/com/merxury/blocker/core/ui/topBar/SelectedAppTopBar.kt
@@ -14,40 +14,28 @@
  * limitations under the License.
  */
 
-package com.merxury.blocker.feature.search.component
+package com.merxury.blocker.core.ui.topBar
 
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextFieldDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.pluralStringResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.ImeAction
-import androidx.compose.ui.text.input.TextFieldValue
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import com.merxury.blocker.core.designsystem.component.BlockerConfirmAlertDialog
 import com.merxury.blocker.core.designsystem.component.BlockerMediumTopAppBar
-import com.merxury.blocker.core.designsystem.component.BlockerSearchTextField
-import com.merxury.blocker.core.designsystem.component.BlockerTopAppBar
 import com.merxury.blocker.core.designsystem.icon.BlockerActionIcon
 import com.merxury.blocker.core.designsystem.icon.BlockerIcons
 import com.merxury.blocker.core.designsystem.theme.BlockerTheme
-import com.merxury.blocker.feature.search.R.plurals
-import com.merxury.blocker.feature.search.R.string
+import com.merxury.blocker.core.ui.R
 
 @Composable
 fun SelectedAppTopBar(
-    selectedAppCount: Int,
+    title: Int,
+    selectedItemCount: Int,
     selectedComponentCount: Int,
     onNavigationClick: () -> Unit,
     onSelectAll: () -> Unit,
@@ -62,9 +50,9 @@ fun SelectedAppTopBar(
     }
     BlockerMediumTopAppBar(
         title = pluralStringResource(
-            id = plurals.feature_search_selected_app_count,
-            count = selectedAppCount,
-            selectedAppCount,
+            id = title,
+            count = selectedItemCount,
+            selectedItemCount,
         ),
         navigation = {
             IconButton(onClick = onNavigationClick) {
@@ -98,7 +86,7 @@ fun SelectedAppTopBar(
     if (showBlockAllDialog) {
         BlockerConfirmAlertDialog(
             text = pluralStringResource(
-                id = plurals.feature_search_block_all,
+                id = R.plurals.core_ui_block_all,
                 count = selectedComponentCount,
                 selectedComponentCount,
             ),
@@ -109,7 +97,7 @@ fun SelectedAppTopBar(
     if (showEnableAllDialog) {
         BlockerConfirmAlertDialog(
             text = pluralStringResource(
-                id = plurals.feature_search_enable_all,
+                id = R.plurals.core_ui_enable_all,
                 count = selectedComponentCount,
                 selectedComponentCount,
             ),
@@ -120,45 +108,13 @@ fun SelectedAppTopBar(
 }
 
 @Composable
-fun SearchBar(
-    modifier: Modifier = Modifier,
-    keyword: TextFieldValue,
-    onSearchTextChanged: (TextFieldValue) -> Unit,
-    onClearClick: () -> Unit,
-) {
-    val colors = TextFieldDefaults.colors(
-        focusedIndicatorColor = Color.Transparent,
-        unfocusedIndicatorColor = Color.Transparent,
-    )
-    BlockerTopAppBar(
-        title = stringResource(id = string.feature_search_searching),
-        actions = {
-            BlockerSearchTextField(
-                keyword = keyword,
-                placeholder = {
-                    Text(
-                        text = stringResource(id = string.feature_search_search_hint),
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                    )
-                },
-                onValueChange = onSearchTextChanged,
-                onClearClick = onClearClick,
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-                modifier = modifier.fillMaxWidth(),
-                colors = colors,
-            )
-        },
-    )
-}
-
-@Composable
 @Preview
 fun SelectedAppTopBarPreview() {
     BlockerTheme {
         Surface {
             SelectedAppTopBar(
-                selectedAppCount = 3,
+                title = R.plurals.core_ui_selected_app_count,
+                selectedItemCount = 3,
                 selectedComponentCount = 6,
                 onNavigationClick = {},
                 onSelectAll = {},
