+++ core/ui/src/main/res/values-zh-rTW/strings.xml
@@ -55,4 +55,13 @@
     <string name="core_ui_sort_options">排序選項</string>
     <string name="core_ui_collapse_list">折疊列表</string>
     <string name="core_ui_expand_list">展開列表</string>
+    <plurals name="core_ui_block_all">
+        <item quantity="other">此操作將禁用%1$d個組件，是否要繼續？</item>
+    </plurals>
+    <plurals name="core_ui_enable_all">
+        <item quantity="other">此操作將启用%1$d個組件，是否要繼續？</item>
+    </plurals>
+    <plurals name="core_ui_selected_app_count">
+        <item quantity="other">%1$d個應用程式已選取</item>
+    </plurals>
 </resources>
