+++ feature/appdetail/src/main/res/values-zh-rTW/strings.xml
@@ -52,10 +52,4 @@
     <plurals name="feature_appdetail_selected_provider_count">
         <item quantity="other">已選擇%1$d個內容提供者</item>
     </plurals>
-    <plurals name="feature_appdetail_block_all">
-        <item quantity="other">此操作將禁用%1$d個組件，是否要繼續？</item>
-    </plurals>
-    <plurals name="feature_appdetail_enable_all">
-        <item quantity="other">此操作將启用%1$d個組件，是否要繼續？</item>
-    </plurals>
 </resources>
