+++ feature/search/src/main/java/com/merxury/blocker/feature/search/SearchScreen.kt
@@ -65,8 +65,8 @@ import com.merxury.blocker.core.ui.screen.ErrorScreen
 import com.merxury.blocker.core.ui.screen.InitializingScreen
 import com.merxury.blocker.feature.applist.AppListViewModel
 import com.merxury.blocker.feature.search.component.FilteredComponentItem
+import com.merxury.blocker.core.ui.topBar.SelectedAppTopBar
 import com.merxury.blocker.feature.search.component.SearchBar
-import com.merxury.blocker.feature.search.component.SelectedAppTopBar
 import com.merxury.blocker.feature.search.model.ComponentTabUiState
 import com.merxury.blocker.feature.search.model.LocalSearchUiState
 import com.merxury.blocker.feature.search.model.SearchUiState
@@ -228,7 +228,8 @@ fun TopBar(
     ) { targetState ->
         if (targetState) {
             SelectedAppTopBar(
-                selectedAppCount = searchUiState.selectedAppList.size,
+                title = R.plurals.feature_search_selected_app_count,
+                selectedItemCount = searchUiState.selectedAppList.size,
                 selectedComponentCount = searchUiState.selectedComponentList.size,
                 onNavigationClick = onNavigationClick,
                 onSelectAll = onSelectAll,
